requirements.txt
pymongo
requests
pyspark
python-dotenv
numpy